+++ examples/actix_ssr/vite.config.js
@@ -1,10 +1,11 @@
 import react from '@vitejs/plugin-react';
 import laravel from 'laravel-vite-plugin';
 import { defineConfig } from 'vite';
+import tailwindcss from "@tailwindcss/vite"
 
 export default defineConfig(() => {
     return {
-        plugins: [    
+        plugins: [
             laravel({
                 input: 'www/app.tsx',
                 buildDirectory: 'bundle',
@@ -12,6 +13,7 @@ export default defineConfig(() => {
                 ssr: "www/ssr.tsx",
             }),
             react(),
+            tailwindcss()
         ],
         // important to serve statics from public dir directly from "localhost:5173/" instead of "localhost:5173/public"
         // just "public" without slash prefix won't work